Renamed `crossOriginIsolation.js` to `crossOriginIsolatedQuestion.js`
