+++ server.js
@@ -1008,7 +1008,7 @@ module.exports.initExpress = function () {
       res.locals.navSubPage = 'preview';
       next();
     },
-    require('./middlewares/crossOriginIsolation'),
+    require('./middlewares/crossOriginIsolatedQuestion'),
     require('./pages/shared/floatFormatters'),
     require('./pages/instructorQuestionPreview/instructorQuestionPreview'),
   ]);
@@ -1337,7 +1337,7 @@ module.exports.initExpress = function () {
     require('./middlewares/selectAndAuthzInstanceQuestion'),
     // don't use logPageView here, we load it inside the page so it can get the variant_id
     require('./middlewares/studentAssessmentAccess'),
-    require('./middlewares/crossOriginIsolation'),
+    require('./middlewares/crossOriginIsolatedQuestion'),
     require('./pages/studentInstanceQuestionHomework/studentInstanceQuestionHomework'),
     require('./pages/studentInstanceQuestionExam/studentInstanceQuestionExam'),
   ]);
@@ -1447,7 +1447,7 @@ module.exports.initExpress = function () {
       res.locals.navSubPage = 'preview';
       next();
     },
-    require('./middlewares/crossOriginIsolation'),
+    require('./middlewares/crossOriginIsolatedQuestion'),
     require('./pages/shared/floatFormatters'),
     require('./pages/instructorQuestionPreview/instructorQuestionPreview'),
   ]);
